extracted run script enum and builder to separate classes
